First working version of `detectPlatform` for vc6
Figured out how to access vc6 info from project `py-videocore6`.
It's a shame that I can't run this project local, but the
memory mapping info proves itself to be useful.

- Added vc6-specific files under `Lib/VideoCore/vc6/`.
- Adjusted Makefile for this
- Added vc6 core-fields to output of `detectPlatform`

The output for vc6 in `detectPlatform` can now be expanded
as deemed necessary.

NOTE: Not sure if I need `readwrite4.c`, not used yet.
      It might be a python-specific thing.