Add handling of Performance Counter registers
*This was a pull request on the original `QPULib`, which after two years still wasn't merged.*

This enables viewing and initialization of the performance registers in the `RegisterMap`.

These can be used for hardware profiling whern running kernels. Notably, they will be useful when optimizing the generated kernel code.

- A separate class `PerformanceCounters` has been added; this because the handling is pretty extensive and was 'polluting' `RegisterMap`.
- Example usage has been added to `Rot3DLib`.

----
It took me a while to get this to work because the full functionality was not documented. I needed the Errata (#65) to get it to work; a single bit needed to be set for that.