Add SFU write registers to Encoding
*(Text from original PR)*

Initial changes for #43; this is not a complete solution.

This adds the definitions and handling of the four input registers for the Special Function Unit.

In addition, the following changes to `encodeDestReg()` and `encodeSrcReg()`:

- Added extensive header commenting, to document how I understand it
- Added TODOs which I think are relevant
- Reformatted code to make common blocks more visible
- Minor changes in code, mainly to aid reading

Note that this only adds the *definitions* of the special registers; no code is yet generated that actually uses them. That's the next step.

======

*(Text from preceding issue #43)*

Usage of the Special Function Unit (SFU) is something I would really like to add to `QPULib`.

Page 23 of th `VideoCore` Reference tells you how to do it:

- Load a given location (e.g. `SFU_RECIPSQRT`) in the A/B regfile space
- Don't read special register `r4` for two cycles
- Read result from register `r4`

This is simple in principle, but I would like to know the opcodes/commands to move values between different addresses in the regfiles. Would you mind explaining this to me?

The rest I think I can figure out using the `QPULib` DSL.

----

I have to say I'm a bit underwhelmed by the SFU. It appears to deal with one value at a time (as opposed to the regular blocks of 16) and on top of that there is the two-cycle wait. Still, I hope it is of some value.

Accessing the SFU sounds very similar to accessing the TMU. You could follow the gather() and receive() functions through the compiler to see how it works.

WR: I see fields add_a, mul_a etc as bit fields in the Instruction Encoding tabel (page 26).
    Scrolling down, in table 3 on page 28, I see that special register r4 can be selected here for read.

My suggestion to look at `receive()` and `gather()` was not just to see how the DSL syntax construction works.
You will see the AST elements (`LOAD_RECEIVE` and `TMU0_ADDR`) used to represent these operations and, if you grep for these in the `Source` and `Target` dirs, you'll see how the compiler handles them.
These are relevant to the SFU: `gather()` shows how to write to a special register and `receive()` shows how to read from r4.