Refactoring of Liveness
This is to get a grip on what it actually does.
Also, small cleanups and added comments.

Also:

- Fixed `Seq::size()`, see below
- Adjusted preference of regfile A and B per variable for `v3d`.
  There is only one regfile on `v3d`, hence determining preference is useless

--------------------------------------------------------

Here's one for the Hall of Shame, previous definition:

    bool size() const { return numElems; }

How much of an idiot can I be?
I kept reading over it for ages, because how can size be wrong, right?
Hours of confusion have now been explained.