Fixed translation of small immediates
This was an interesting one.
Turns out that small immediates >= 16 are regarded as floats within the QPU's.

Therefore, split conversion of float and int small immediates; for int's you can now
also use powers of the legal values (-16 <= val < 15), by inclusion of `shl`.
This allows for int 16 to be passed in, which occurs commonly.

- Split setting of cond tag for small immediates to separate function. This prob has
  broader usage that just small imm's.

Example `ReqRcv` now finally works as expected (but not as intended!).

Also:

- Code for branch encoding in `v3d::KernelDriver` to separate method (cleanup)
- Added index num offset to pointer deref in `Interpreter`. This is to make the working
  overall consistent between all possible run platforms (which includes emulator and interpreter)
- Added v3d `tmuwt` instructions. This is also added at the intermediate code level, because
  vc4 has no notion of this.
- Cleanup of `InstrTag`, to make adding v3d-specific tags easier and for overall easing of eyeballing.