Edits and fixes while examining working of HeatMap
*Oof, too many changes. Must make smaller commits.*

`HeatMap` doesn't work at various levels, this commit does code cleanup, refactoring and
addition of debug support (which I hope is globally useful).

When inserting/added new elements, the sequence size was incremented multiple times, leading
to errant zero-instructions. Added fix and refactoring to `Seq` for this.

Added debug function `check_zeroes()` to debug this issue.

This is used to indicate no input registers for an instruction, notably for `tidx()` and `eidx()`.
Added handling in `v3d`. It doesn't (shouldn't - assertions will fail) occur in `vc4`.

Tried it, and it *almost* works. Not good enough.
What it does is map registers to instructions that use it. I have thought hard for a use-case
for this, and failed to convince myself I need it. It was disabled in the original code anyway.

These get caught at top-level compilation and stored as compile errors. Compilation is, of course,
abruptly terminated when this happens.

A best-effort attempt is done to write all available representation to file.
If errors occurred, a message is written at the top of the file

- Moved `assign()` to its own include file
- Moved `invoke_qpu()` from Kernel to KernelDriver level
- Lots of comment editing and formatting

- Added debug function to `Liveness` to dump the possible allocations-vector to output
- Refactored `Pretty` for target-level to return strings, instead of having to pass the file pointer
  all the way down
- Redefined some pretty-print instructions as member methods of the structs they print
- Adjusted `v3d` `rotate` instruction generation so that it can take any source instead of just `r0`.
  Done by inserting a move to `r0`.
- Added `vc4` opcode dump function to `mesa`. Not in use yet, needs some more work.