Added passing of dummy uniform value to vc4
Often (not always!), the final param is passed garbled when input as a uniform to
a kernel program executing on vc4 hardware. It appears to happen to direct Float/Int values.
After spending days on this with paranoid debugging, I could not find the cause and gave up.

Instead, I'll just pass a final dummy uniform value,which can be mangled to
the heart's content of the hardware ('just' - hehe, wiseass).

Example `HeatMap` is now in an interim but working state, commiting it anyway.

Also:

- Refactored compilation in `KernelDriver` to better split `vc4` and `v3d` compilation.
  The ultimate goal is to bring the compilation under the device-dependent kernel drivers.
- Added dump-method to `Common/SharedArray` for debugging
- Small refactorings in `Translate`, to make better sense of the code
- Refactoring and extensive commenting, to clarify the issue solved here, in `vc4/Invoke`