Added booleans for If/When
In Translate, conditions are now stored in implicit boolean var's for
translation of conditions in If and When.

This is to deal with the differing approaches of condition flag handling
in v3d and vc4.
It's a bit more inefficient, but it does the job.

Autotest now works, needs major code cleanup.
Other unit tests, notably for v3d, still need to be checked.