Setting of square matrix dimension with decorator
A decorator is function is used to set the sizes of matrixes for multiplication.
This replaces the previous attempt using templates.

The big advantage of this is that the dimension can now be set a runtime (at kernel compile).
The ugly bit is that it uses a private global integer within the Matrix kernel code to set it.

I can live with this for now, until a better implementation presents itself.
For now, the only thing to avoid is kernel-compiling matrix multiplication on parallel threads.
And we don't use or supports multiple threads yet.