Removed store request operation
`store(val, address)` is practically identical to the usual `*address = val`;
the only difference is where it waits for a previous store to complete.

I believe the original idea was to optimize the store for multiple writes,
but after research I conclude that this is not possible.
I don't want to support both write syntaxes, and I prefer the *-version.
Therefore, `store()` hereby officially dies.

Also:

- Cleaned up and refactored Ptr