Cleaned up and optimized v3d code generation
This is the first serious attempt at optimization of v3d code.

The Rot3D kernel 3 has been adjusted with all the stuff below.
Truth be told, the results are underwhelming; it still looks like Rot3D is IO-bound.

* Gather limit set to 8 for v3d
This is different vc4, with has gather limit 4 (needs to be documented)

* Added increment method for sequential pointers
This defines the increment (64) *once*, and is added directly to the pointer value.
This is to avoid creating '64' from scratch every time, which has a larger code overhead than you would expect.

* Disabled code adjustment during assembly for v3d
Some things that are required for vc4 work on v3d, but are not required and just add overhead.

* Added special handling of instructions during v3d assembly
The framework for this has now been added.
In particular: a fetch using TMU can now be combined with a sequential pointer increment into
one instruction.

This is the first actual case in V3DLib where the add and mul ALUs operate in parallel.