Remove stride by QPU ID for uniform pointers
Initializing by default a stride to uniform pointers with the QPU index turned out to be a bad idea.
This is only useful in the simplest of situations, it obstructs kernels doing anything non-trivial.
Thus removing it.

Woopie, a breaking API change. Both my users are going to kill me.
This one led to Examples having to be fixed. Upping the minor version.

Also code cleanup as I encountered it.