DFT in unit test now functional
This is the most basic implementation you can think of, just complex array
multiplication. No effort to optimize at all, above whatever is optimized
in 'standard' array mult kernel.

The issues were exactly what you would expect when doing array mult R=AxB, but keep
on forgetting that the B matrix needs to be passed transposed to the kernel.

Also, output column size needs to be multiple of 16, which is a bummer
when dealing with 1D arrays.

Now I can breathe again.