Fixed liveness of conditionally assigned variables
Fix for this special case, if this is the *only* usage of `A6`:

    where ZC: LI A6 <- 1
    where ZC: A5 <- or(A6, A6)

... `A6` is considered live from the start of the program
onward till the last (and only) usage.

This unnecessarily ties up a register for a long duration, complicating the
allocation by creating a false shortage of registers.

Unit tests passing for conditions on vc4. Some work still needed for other tests, notably Matrix.

For gained insights, read comments of 'class Liveness' and method 'useDefReg()'.

- Added class UsageReg, which keeps track variable usage within a given program.
- Restructured liveness to get this working
- Added loads of documentation in commenting for Liveness
- Added method 'Seq<>::remove_set()', fix indentation in Seq

(You are looking at the result of two man-days of research)