Started work on combing of immediates
Interim commit, not done yet.

Initializing immediates for v3d can be costly because of the limited range for small immediates.
It is therefore a good idea to combine inits for the same values.

The issue here is to avoid reusing immediate assignments (in target lang) which
are in conditional blocks.

If this doesn't make any sense to you, I understand. It took me the full weekend
to wrap my brain around this issue.