Added subsitution of same LI's
The substitution is the first step in the optimization.

The goal is to minimize the number non-basic immediates, notably for v3d, which require assembly.
The idea, however, could be expanded to basic immediates in vc4 that get used a lot (e.g.
value 64 for ptr increment).

It needs to be done as first, in order not to screw up further logic downstream.
This is because the instructions are adjusted (and removed!) due to substitution, so
the liveness needs to be redone.

This needed a lot of scaffolding and framework rewriting:

- Added logic for detecting if conditional blocks allow the substition
- Added subsitution of instructions with NOP
- Added plenty of checks if previous are allowed

Also refactoring:

- Moved CFG to within Liveness (where is belongs)
- UseDefReg() now defined as method of UseDefReg: UserDefReg::set_used()
- Small comment and code cleanups