Completed conversion of statement sequences to arrays
Top-level sequences are now implemented as arrays.
Notably, the statement stack uses arrays.

To reiterate, previously in the AST, sequences where stored in tree-form, e.g:

    (SET SKIP, (SET stmt1, (SET stmt2, (SET,,,,,,))))

The issue with this was that for really long sequences, the stack overflowed due to deep recursion
when interpreting or compiling.
This does not happen any more (although for giant kernels, the compile time is painfully long).

This commit fixes all lingering issues, the feature hereby goes live.

- All unit tests passing.
- Some performance improvement fixes for Liveness
- Minor code cleanup for unit tests