Changes due to examination of conditional execution
The issue here is that the cond assign and push is entirely separate
for add and mul alus.

This raises issues when combining add/mul instructions, because the code
assumes there is only one set of assign/push tags per instruction.

The easiest way out is to always keep the conditional handling in the add alu,
but of course we're pig-headed about this. It should work!

v3d only. This is actually a commit before reset, gpu hanging due to failing tests.

- Moved setting of cond assign and push tags to V3DLib::Instr
- Minor refactorings for better understanding
- Added comments about gained insights