Refactored and consolidated matrix multiplication code
Module Matrix has been revised so that the same code is used for
Float and Complex matrix multiplication.

This means that a lot of stuff has been rewritten as templates.
It cuts down on the amount of parallel code but increases the complexity.
I hope my brain can handle it downstream.

Bumped up used c++ standard version to c++17 for this, because this implementation
requires 'if constexpr' expressions.