Added block matrix multiplication for complex matrices
This one made me a bit queasy.
Unit tests were failing and I was convinced that the issue was with the added code.
So I literally days debugging to find the issue without any success.

Turns out the actual problem was missing file dependencies for the unit tests.
So, changes in the Matrix multiplication code were not reflected in the tests, leading
to unexpected results.

I should be learning a lesson here, hope it sticks.