All unit tests passing for vc4
This is not so much a fix, but an observation that DFT on vc4 is really imprecise.
The cumulative error due to repeated float mult is horrendous.

On the plus side, using phase addition instead of resulting sincos multiplication
nicely improved the precision. Adjusted expected precision where relevant in unit tests.

- Added comment on cumulative error issue in relevant unit test
- Fixed DFT unit test to work around the imprecision. Yes, this is cheating
- Removed some debug output, some code cleanup